Move download functions into LiveJournalDL class
- This lets us login to LiveJournal only if needed, and only
  once per run.
- I felt it was cleaner to merge download_posts.py and
  download_comments.py into one file: livejournaldl.py
- No longer need auth.py -- configuration has been moved to
  config.json and remaining details moved to where they're
  used.
